backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.121.3",
    "pydantic-ai>=1.21.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "psycopg[binary]>=3.2.0",
    "xgboost>=3.1.2",
    "supabase>=2.24.0",
    "joblib>=1.5.2",
    "scikit-learn>=1.7.2",
    "lightgbm>=4.6.0",
    "prophet>=1.2.1",
    "plotly>=6.5.0",
    "pandas>=2.0.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "llama-index-core>=0.10.0",
    "llama-index-embeddings-huggingface>=0.1.0",
    "llama-index-vector-stores-chroma>=0.1.0",
    "llama-index-readers-file>=0.1.0",
    "llama-index-llms-gemini>=0.1.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "python-multipart>=0.0.9",
]
